+++ benchmark/Streamly/Benchmark/FileSystem/Handle/ReadWrite.hs
@@ -36,7 +36,7 @@ import qualified Streamly.Internal.FileSystem.Handle as IFH
 import qualified Streamly.Internal.Memory.ArrayStream as AS
 import qualified Streamly.Internal.Unicode.Array.Char as IUA
 import qualified Streamly.Internal.Data.Stream.IsStream as IP
-import qualified Streamly.Memory.Array as A
+import qualified Streamly.Data.Array.Storable.Foreign as A
 import qualified Streamly.Prelude as S
 
 import Gauge hiding (env)