+++ test/String.hs
@@ -9,7 +9,7 @@ import Test.QuickCheck.Monadic (run, monadicIO, assert)
 
 import           Test.Hspec as H
 
-import qualified Streamly.Memory.Array as A
+import qualified Streamly.Data.Array.Storable.Foreign as A
 import qualified Streamly.Internal.Memory.ArrayStream as AS
 import qualified Streamly.Prelude as S
 import qualified Streamly.Unicode.Stream as SS